separate database class
